Moved EOM to a seals.py and added DBSig message there too
